Simplify lib architecture and fix cascade deletes.
Unify user app-data cleanup, collapse shallow MCQ/auth/analytics splits, honest agent-discovery metadata, and extract question-card fetch/DOM helpers.
