+++ src/lib/auth/urls.ts
@@ -1,16 +1,7 @@
-import { dev } from '$app/environment';
 import { resolve } from '$app/paths';
-import { env as publicEnv } from '$env/dynamic/public';
+import { getSiteUrl } from '$lib/site-url';
 
-export const PRODUCTION_SITE_URL = 'https://freeappractice.org';
-
-/** Canonical site origin for sitemaps, emails, and structured data. */
-export function getSiteUrl(requestOrigin?: string): string {
-	const configured = publicEnv.PUBLIC_BASE_URL?.replace(/\/+$/, '');
-	if (configured) return configured;
-	if (dev) return (requestOrigin ?? 'http://localhost:5173').replace(/\/+$/, '');
-	return PRODUCTION_SITE_URL;
-}
+export { getSiteUrl, PRODUCTION_SITE_URL } from '$lib/site-url';
 
 type AuthCallbackPath =
 	| '/app'
