+++ src/lib/client/activation-analytics.ts
@@ -1,12 +1,54 @@
 import { hasAnalyticsConsent } from '$lib/client/analytics-consent';
-import {
-	daysBetweenCalendarDays,
-	latencyBucket,
-	localCalendarDay,
-	type QuestionFailureKind,
-	type QuestionSource
-} from '$lib/client/activation-funnel-metrics';
 import { capturePostHogEvent } from '$lib/client/posthog-analytics';
+import {
+	classifyQuestionFailureFromStatus,
+	type QuestionFailureKind
+} from '$lib/question-failure';
+
+export type { QuestionFailureKind };
+export { classifyQuestionFailureFromStatus as classifyQuestionFailure };
+
+export type LatencyBucket = '0-500ms' | '500-1000ms' | '1-2s' | '2-5s' | '5s+';
+export type QuestionSource = 'cached' | 'generated';
+
+export function latencyBucket(ms: number): LatencyBucket {
+	if (!Number.isFinite(ms) || ms < 0) return '5s+';
+	if (ms < 500) return '0-500ms';
+	if (ms < 1000) return '500-1000ms';
+	if (ms < 2000) return '1-2s';
+	if (ms < 5000) return '2-5s';
+	return '5s+';
+}
+
+export class QuestionRequestError extends Error {
+	readonly status: number | null;
+	readonly failureKind: QuestionFailureKind;
+
+	constructor(message: string, status: number | null) {
+		super(message);
+		this.name = 'QuestionRequestError';
+		this.status = status;
+		this.failureKind = classifyQuestionFailureFromStatus(status);
+	}
+}
+
+export function questionSourceFromCachedFlag(cached: unknown): QuestionSource {
+	return cached === true ? 'cached' : 'generated';
+}
+
+export function localCalendarDay(date = new Date()): string {
+	const y = date.getFullYear();
+	const m = String(date.getMonth() + 1).padStart(2, '0');
+	const d = String(date.getDate()).padStart(2, '0');
+	return `${y}-${m}-${d}`;
+}
+
+export function daysBetweenCalendarDays(earlier: string, later: string): number {
+	const a = Date.parse(`${earlier}T00:00:00`);
+	const b = Date.parse(`${later}T00:00:00`);
+	if (!Number.isFinite(a) || !Number.isFinite(b) || b < a) return 0;
+	return Math.round((b - a) / 86_400_000);
+}
 
 /** localStorage key for the anonymous activation journey (consent-gated). */
 export const ACTIVATION_JOURNEY_KEY = 'ph_activation_journey_key';
