--- src/lib/components/fancy/scroll-and-swap/index.ts
@@ -1,3 +1,0 @@
-import ScrollAndSwap from "./scroll-and-swap.svelte";
-
-export { ScrollAndSwap };
