--- src/lib/components/magic/stagger-text/index.ts
@@ -1,1 +1,0 @@
-export { default as StaggerText } from "./stagger-text.svelte";
