+++ src/lib/components/questions/question-shell.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
-	import { captureGenerateClicked } from '$lib/client/activation-analytics';
-	import QuestionCard from '$lib/components/questions/question-card.svelte';
-	import QuestionSelector from '$lib/components/questions/question-selector.svelte';
+	import PracticeShell from '$lib/components/practice/practice-shell.svelte';
 	import type { QuestionCardProps } from '$lib/questions/types';
 
+	/** Public MCQ-only practice chrome — thin wrapper over PracticeShell. */
 	type QuestionShellProps = {
 		selectedClass?: string;
 		selectedUnit?: string;
@@ -19,37 +18,15 @@
 		selectedUnit = $bindable(''),
 		unitRange = $bindable<number[] | undefined>(undefined),
 		requestVersion = $bindable(0),
-		generateLabel,
-		onGenerate,
-		onSelectionChange,
-		...cardProps
+		...rest
 	}: QuestionShellProps = $props();
-
-	function handleSelectionChange(className: string, unit: string): void {
-		requestVersion = 0;
-		onSelectionChange?.(className, unit);
-	}
-
-	function handleGenerate(): void {
-		if (selectedClass) captureGenerateClicked(selectedClass, selectedUnit);
-		requestVersion += 1;
-		onGenerate?.();
-	}
 </script>
 
-<div class="mx-auto mb-8 max-w-5xl">
-	<QuestionSelector
-		bind:selectedClass
-		bind:selectedUnit
-		bind:unitRange
-		{generateLabel}
-		onSelectionChange={handleSelectionChange}
-		onGenerate={handleGenerate}
-	/>
-</div>
-
-<div class="mx-auto min-h-40 max-w-6xl">
-	{#key `${selectedClass}:${selectedUnit}:${requestVersion}`}
-		<QuestionCard {selectedClass} {selectedUnit} {unitRange} {requestVersion} {...cardProps} />
-	{/key}
-</div>
+<PracticeShell
+	bind:selectedClass
+	bind:selectedUnit
+	bind:unitRange
+	bind:requestVersion
+	allowFrq={false}
+	{...rest}
+/>
