--- src/lib/components/spell/slide-up-text/index.ts
@@ -1,2 +1,0 @@
-import SlideUpText from "./slide-up-text.svelte";
-export { SlideUpText };
