+++ src/lib/practice/assign-variant.server.ts
@@ -1,11 +1,21 @@
+import { createHash } from 'node:crypto';
 import type { PracticeExperimentAssignment, PracticeVariant } from '$lib/practice/multi-attempt';
 import {
 	MULTI_ATTEMPT_EXPERIMENT_KEY,
 	MULTI_ATTEMPT_EXPERIMENT_VERSION
 } from '$lib/practice/multi-attempt';
-import { assignPracticeVariant } from '$lib/practice/assign-practice-variant.server';
 import { findUserProfileOrFail } from '$lib/users/profile.server';
 import { isMultiAttemptExperimentEnabled } from '$lib/flags';
+
+/** Stable 50/50 assignment from user id + experiment identity (not client-chosen). */
+export function assignPracticeVariant(
+	userId: string,
+	key = MULTI_ATTEMPT_EXPERIMENT_KEY,
+	version = MULTI_ATTEMPT_EXPERIMENT_VERSION
+): PracticeVariant {
+	const digest = createHash('sha256').update(`${key}:v${version}:${userId}`).digest();
+	return digest[0]! % 2 === 0 ? 'control' : 'multi_attempt_hints';
+}
 
 /** Get or create sticky assignment for the multi-attempt experiment. */
 export async function getOrAssignMultiAttemptVariant(
