+++ src/lib/server/agent-discovery/oauth.ts
@@ -1,27 +1,26 @@
 import { absoluteUrl } from './site';
 
+/**
+ * Auth discovery for agents. This is not a standards-compliant OAuth2 AS —
+ * Free AP Practice uses Better Auth (session cookies + email/social sign-in).
+ * Endpoints below point at real Better Auth / site routes only.
+ */
 export function buildOAuthAuthorizationServerMetadata(requestUrl?: URL) {
 	const issuer = absoluteUrl('/', requestUrl);
 
 	return {
 		issuer,
-		authorization_endpoint: absoluteUrl('/api/auth/sign-in/social/google', requestUrl),
-		token_endpoint: absoluteUrl('/api/auth/get-session', requestUrl),
-		jwks_uri: absoluteUrl('/api/auth/jwks', requestUrl),
-		registration_endpoint: absoluteUrl('/api/auth/sign-up/email', requestUrl),
-		scopes_supported: ['openid', 'profile', 'email', 'offline_access'],
+		authorization_endpoint: absoluteUrl('/login', requestUrl),
+		registration_endpoint: absoluteUrl('/signup', requestUrl),
+		scopes_supported: ['profile', 'email'],
 		response_types_supported: ['code'],
-		grant_types_supported: ['authorization_code', 'refresh_token'],
-		token_endpoint_auth_methods_supported: ['client_secret_post', 'none'],
-		code_challenge_methods_supported: ['S256'],
+		grant_types_supported: [] as string[],
+		token_endpoint_auth_methods_supported: ['none'],
+		service_documentation: absoluteUrl('/auth.md', requestUrl),
 		agent_auth: {
 			skill: absoluteUrl('/auth.md', requestUrl),
 			register_uri: absoluteUrl('/api/auth/sign-up/email', requestUrl),
-			identity_types_supported: ['anonymous', 'verified_email'],
-			anonymous: {
-				credential_types_supported: ['session_cookie'],
-				claim_uri: absoluteUrl('/signup', requestUrl)
-			},
+			identity_types_supported: ['verified_email'],
 			identity_assertion: {
 				assertion_types_supported: ['verified_email'],
 				credential_types_supported: ['session_cookie'],
@@ -38,8 +37,8 @@ export function buildOAuthProtectedResourceMetadata(requestUrl?: URL) {
 	return {
 		resource,
 		authorization_servers: [issuer],
-		scopes_supported: ['profile', 'email', 'offline_access'],
-		bearer_methods_supported: ['header', 'cookie'],
+		scopes_supported: ['profile', 'email'],
+		bearer_methods_supported: ['cookie'],
 		resource_documentation: absoluteUrl('/llms.txt', requestUrl)
 	};
 }
